Add not found exception

+++ app/Actions/UserTaste/DeleteCustomTasteAction.php
@@ -2,6 +2,7 @@
 
 namespace Hedonist\Actions\UserTaste;
 
+use Hedonist\Exceptions\User\CustomTasteNotFoundException;
 use Hedonist\Repositories\User\TasteRepositoryInterface;
 use Illuminate\Support\Facades\Auth;
 
@@ -16,6 +17,10 @@ class DeleteCustomTasteAction
 
     public function execute(DeleteCustomTasteRequest $deleteCustomTasteRequest)
     {
+        $customTaste = $this->tasteRepository->getCustomById($deleteCustomTasteRequest->getCustomTasteId());
+        if (!$customTaste) {
+            throw new CustomTasteNotFoundException('Custom taste not found!');
+        }
         $this->tasteRepository->deleteCustomById($deleteCustomTasteRequest->getCustomTasteId());
     }
 }
